refactor(api): 拆分 locations.ts 和 users.ts 为模块化结构
任务 #73：locations.ts (696行) 和 users.ts (707行) 模块化拆分

参照 teams/ 模式拆为 index.ts + queries.ts + mutations.ts + utils.ts：

locations/
- index.ts (26行): 路由注册 + 导出 locationsRoute
- queries.ts (462行): GET /, /:id, /:id/pois, /:id/tags, /search
- mutations.ts (217行): POST /, PUT /, DELETE /:id, PUT /:id/tags, /:id/pois
- utils.ts (33行): safeJsonParse, requireAdmin

users/
- index.ts (23行): 路由注册 + 导出 usersRoute
- queries.ts (465行): GET /, /pending-approvals, /applications, /teams/joined, /created-teams, /:id
- mutations.ts (94行): PATCH /update
- utils.ts (150行): formatBeijingDateTime, sanitizeUser, validateUserExtra, getUserStats, getUserOngoingTeams

所有文件 <500 行。纯重构，不改功能。
原导入路径兼容：api/src/index.ts 仍 import { locationsRoute } from './routes/locations' 等（解析到目录 index.ts）。

验证：tsc 通过；api lint 0 错误；178 测试全过
