+++ api/src/routes/locations/utils.ts
@@ -1,0 +1,33 @@
+import { eq } from "drizzle-orm";
+import { createAuth } from "../../lib/auth";
+import { createDb } from "../../db";
+import * as schema from "../../db/schema";
+import type { Env } from "../../lib/auth";
+
+/** 安全解析 JSON 字段 */
+export function safeJsonParse<T>(value: string | null | undefined, fallback: T): T {
+  if (!value) return fallback;
+  try {
+    return JSON.parse(value);
+  } catch {
+    if (Array.isArray(fallback)) {
+      return value.split(/[,、]/).map((s: string) => s.trim()).filter(Boolean) as unknown as T;
+    }
+    return fallback;
+  }
+}
+
+/** 验证管理员权限 */
+export async function requireAdmin(c: { env: Env; req: { raw: Request } }) {
+  const authInstance = createAuth(c.env);
+  const session = await authInstance.api.getSession({ headers: c.req.raw.headers });
+  if (!session) throw new Error("未登录");
+  const db = createDb(c.env.DB);
+  const user = await db
+    .select({ role: schema.users.role })
+    .from(schema.users)
+    .where(eq(schema.users.id, session.user.id))
+    .then((rows) => rows[0]);
+  if (!user || user.role !== "admin") throw new Error("无权限访问");
+  return session;
+}
